+++ smos-sync-client/src/Smos/Sync/Client/Command/Login.hs
@@ -2,6 +2,7 @@
 
 module Smos.Sync.Client.Command.Login where
 
+import Control.Monad.Logger
 import Smos.Client
 import Smos.Sync.Client.Env
 import Smos.Sync.Client.OptParse
@@ -10,4 +11,6 @@ loginSmosSyncClient :: Settings -> LoginSettings -> IO ()
 loginSmosSyncClient Settings {..} LoginSettings =
   withClientEnv setServerUrl $ \cenv ->
     withClientVersionCheck cenv $
-      withLogin cenv setSessionPath setUsername setPassword (const $ pure ())
+      runStderrLoggingT $
+        filterLogger (\_ ll -> ll >= setLogLevel) $
+          withLogin cenv setSessionPath setUsername setPassword (const $ pure ())